сохранение последнего местоположения персонажа

+++ server/src/main/kotlin/mysqlEntities/CharacterEntity.kt
@@ -2,9 +2,9 @@ package online.fivem.server.mysqlEntities
 
 class CharacterEntity(
 	val id: Int,
-	val coord_x: Float,
-	val coord_y: Float,
-	val coord_z: Float,
+	val coord_x: Double,
+	val coord_y: Double,
+	val coord_z: Double,
 	val coord_rotation: Float,
 	val pedestrian: Int
 )
